+++ SniffCore.Layouting/ElementStartPosition.cs
@@ -1,0 +1,33 @@
+﻿//
+// Copyright (c) David Wendland. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+//
+
+namespace SniffCore.Layouting
+{
+    /// <summary>
+    ///     Defines the position where the elements has to start in the <see cref="EllipsePanel" />.
+    /// </summary>
+    public enum ElementStartPosition
+    {
+        /// <summary>
+        ///     The first item in the ellipse panel is starting on the left side.
+        /// </summary>
+        Left,
+
+        /// <summary>
+        ///     The first item in the ellipse panel is starting on top.
+        /// </summary>
+        Top,
+
+        /// <summary>
+        ///     The first item in the ellipse panel is starting on the right side.
+        /// </summary>
+        Right,
+
+        /// <summary>
+        ///     The first item in the ellipse panel is starting on bottom.
+        /// </summary>
+        Bottom
+    }
+}